feat phaser entities - created directory for entitiess, minor cleanup in building and grass entitites
